Add self-contained Python script for searching for artwork
This reads in data from STDIN, parses them for the song name, artist, and album,
and then runs the API search and outputs filepaths of downloaded data to
a temporary folder.

This is all managed with one Automator service, which allows you to
automatically add the album artwork by selecting songs in iTunes.